Struct aws_sdk_wellarchitected::client::fluent_builders::CreateWorkload

pub struct CreateWorkload<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateWorkload<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateWorkloadOutput, SdkError<CreateWorkloadError>> where
R::Policy: SmithyRetryPolicy<CreateWorkloadInputOperationOutputAlias, CreateWorkloadOutput, CreateWorkloadError, CreateWorkloadInputOperationRetryAlias>,

pub fn workload_name(self, inp: impl Into<String>) -> Self

The name of the workload.

The name must be unique within an account within a Region. Spaces and capitalization are ignored when checking for uniqueness.

pub fn set_workload_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description for the workload.

pub fn set_description(self, input: Option<String>) -> Self

pub fn environment(self, inp: WorkloadEnvironment) -> Self

The environment for the workload.

pub fn set_environment(self, input: Option<WorkloadEnvironment>) -> Self

pub fn account_ids(self, inp: impl Into<String>) -> Self

Appends an item to AccountIds.

To override the contents of this collection use set_account_ids.

The list of AWS account IDs associated with the workload.

pub fn set_account_ids(self, input: Option<Vec<String>>) -> Self

pub fn aws_regions(self, inp: impl Into<String>) -> Self

Appends an item to AwsRegions.

To override the contents of this collection use set_aws_regions.

The list of AWS Regions associated with the workload, for example, us-east-2, or ca-central-1.

pub fn set_aws_regions(self, input: Option<Vec<String>>) -> Self

pub fn non_aws_regions(self, inp: impl Into<String>) -> Self

Appends an item to NonAwsRegions.

To override the contents of this collection use set_non_aws_regions.

The list of non-AWS Regions associated with the workload.

pub fn set_non_aws_regions(self, input: Option<Vec<String>>) -> Self

pub fn pillar_priorities(self, inp: impl Into<String>) -> Self

Appends an item to PillarPriorities.

To override the contents of this collection use set_pillar_priorities.

The priorities of the pillars, which are used to order items in the improvement plan. Each pillar is represented by its PillarReviewSummary$PillarId.

pub fn set_pillar_priorities(self, input: Option<Vec<String>>) -> Self

pub fn architectural_design(self, inp: impl Into<String>) -> Self

The URL of the architectural design for the workload.

pub fn set_architectural_design(self, input: Option<String>) -> Self

pub fn review_owner(self, inp: impl Into<String>) -> Self

The review owner of the workload. The name, email address, or identifier for the primary group or individual that owns the workload review process.

pub fn set_review_owner(self, input: Option<String>) -> Self

pub fn industry_type(self, inp: impl Into<String>) -> Self

The industry type for the workload.

If specified, must be one of the following:

Agriculture

Automobile

Defense

Design and Engineering

Digital Advertising

Education

Environmental Protection

Financial Services

Gaming

General Public Services

Healthcare

Hospitality

InfoTech

Justice and Public Safety

Life Sciences

Manufacturing

Media & Entertainment

Mining & Resources

Oil & Gas

Power & Utilities

Professional Services

Real Estate & Construction

Retail & Wholesale

Social Protection

Telecommunications

Travel, Transportation & Logistics

Other

pub fn set_industry_type(self, input: Option<String>) -> Self

pub fn industry(self, inp: impl Into<String>) -> Self

The industry for the workload.

pub fn set_industry(self, input: Option<String>) -> Self

pub fn lenses(self, inp: impl Into<String>) -> Self

Appends an item to Lenses.

To override the contents of this collection use set_lenses.

The list of lenses associated with the workload. Each lens is identified by its LensSummary$LensAlias.

pub fn set_lenses(self, input: Option<Vec<String>>) -> Self

pub fn notes(self, inp: impl Into<String>) -> Self

The notes associated with the workload.

pub fn set_notes(self, input: Option<String>) -> Self

pub fn client_request_token(self, inp: impl Into<String>) -> Self

A unique case-sensitive string used to ensure that this request is idempotent (executes only once).

You should not reuse the same token for other requests. If you retry a request with the same client request token and the same parameters after it has completed successfully, the result of the original request is returned.

This token is listed as required, however, if you do not specify it, the AWS SDKs automatically generate one for you. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the request will fail.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

The tags to be associated with the workload.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self