Module aws_sdk_wellarchitected::model

Modules

answer

See Answer

answer_summary

See AnswerSummary

choice

See Choice

choice_answer

See ChoiceAnswer

choice_answer_summary

See ChoiceAnswerSummary

choice_update

See ChoiceUpdate

improvement_summary

See ImprovementSummary

lens_review

See LensReview

lens_review_report

See LensReviewReport

lens_review_summary

See LensReviewSummary

lens_summary

See LensSummary

lens_upgrade_summary

See LensUpgradeSummary

milestone

See Milestone

milestone_summary

See MilestoneSummary

notification_summary

See NotificationSummary

pillar_difference

See PillarDifference

pillar_review_summary

See PillarReviewSummary

question_difference

See QuestionDifference

share_invitation

See ShareInvitation

share_invitation_summary

See ShareInvitationSummary

validation_exception_field

See ValidationExceptionField

version_differences

See VersionDifferences

workload

See Workload

workload_share

See WorkloadShare

workload_share_summary

See WorkloadShareSummary

workload_summary

See WorkloadSummary

Structs

Answer

An answer of the question.

AnswerSummary

An answer summary of a lens review in a workload.

Choice

A choice available to answer question.

ChoiceAnswer

A choice that has been answered on a question in your workload.

ChoiceAnswerSummary

A choice summary that has been answered on a question in your workload.

ChoiceUpdate

A list of choices to be updated.

ImprovementSummary

An improvement summary of a lens review in a workload.

LensReview

A lens review of a question.

LensReviewReport

A report of a lens review.

LensReviewSummary

A lens review summary of a workload.

LensSummary

A lens summary of a lens.

LensUpgradeSummary

Lens upgrade summary return object.

Milestone

A milestone return object.

MilestoneSummary

A milestone summary return object.

NotificationSummary

A notification summary return object.

PillarDifference

A pillar difference return object.

PillarReviewSummary

A pillar review summary of a lens review.

QuestionDifference

A question difference return object.

ShareInvitation

The share invitation.

ShareInvitationSummary

A share invitation summary return object.

ValidationExceptionField

Stores information about a field passed inside a request that resulted in an exception.

VersionDifferences

The differences between the base and latest versions of the lens.

Workload

A workload return object.

WorkloadShare

A workload share return object.

WorkloadShareSummary

A workload share summary return object.

WorkloadSummary

A workload summary return object.

Enums

AnswerReason

ChoiceReason

ChoiceStatus

DifferenceStatus

LensStatus

NotificationType

PermissionType

Permission granted on a workload share.

Risk

The risk for a given workload, lens review, pillar, or question.

ShareInvitationAction

Share invitation action taken by contributor.

ShareStatus

The status of a workload share.

ValidationExceptionReason

The reason why the request failed validation.

WorkloadEnvironment

The environment for the workload.

WorkloadImprovementStatus

The improvement status for a workload.