Module aws_sdk_wellarchitected::operation

Structs

AssociateLenses

Associate a lens to a workload.

CreateMilestone

Create a milestone for an existing workload.

CreateWorkload

Create a new workload.

CreateWorkloadShare

Create a workload share.

DeleteWorkload

Delete an existing workload.

DeleteWorkloadShare

Delete a workload share.

DisassociateLenses

Disassociate a lens from a workload.

GetAnswer

Get the answer to a specific question in a workload review.

GetLensReview

Get lens review.

GetLensReviewReport

Get lens review report.

GetLensVersionDifference

Get lens version differences.

GetMilestone

Get a milestone for an existing workload.

GetWorkload

Get an existing workload.

ListAnswers

List of answers.

ListLensReviewImprovements

List lens review improvements.

ListLensReviews

List lens reviews.

ListLenses

List the available lenses.

ListMilestones

List all milestones for an existing workload.

ListNotifications

List lens notifications.

ListShareInvitations

List the workload invitations.

ListTagsForResource

List the tags for a resource.

ListWorkloadShares

List the workload shares associated with the workload.

ListWorkloads

List workloads. Paginated.

TagResource

Adds one or more tags to the specified resource.

UntagResource

Deletes specified tags from a resource.

UpdateAnswer

Update the answer to a specific question in a workload review.

UpdateLensReview

Update lens review.

UpdateShareInvitation

Update a workload invitation.

UpdateWorkload

Update an existing workload.

UpdateWorkloadShare

Update a workload share.

UpgradeLensReview

Upgrade lens review.